+++ assets/src/components/activities/multi_input/sections/MultiInputStem.tsx
@@ -4,7 +4,7 @@ import { MultiInputSchema } from 'components/activities/multi_input/schema';
 import { InputRefToolbar } from 'components/activities/multi_input/sections/InputRefToolbar';
 import { RichTextEditorConnected } from 'components/content/RichTextEditor';
 import { CommandContext } from 'components/editing/elements/commands/interfaces';
-import { elementsRemoved } from 'components/editing/utils';
+import { elementsRemoved } from 'components/editing/slateUtils';
 import { InputRef } from 'data/content/model/elements/types';
 import React from 'react';
 import { Editor } from 'slate';
