+++ assets/src/components/editing/elements/blockcode/codeblockActions.tsx
@@ -1,5 +1,5 @@
 import { Editor, Element, Transforms } from 'slate';
-import { isActive } from '../../utils';
+import { isActive } from '../../slateUtils';
 import { createButtonCommandDesc } from 'components/editing/elements/commands/commandFactories';
 import { Model } from 'data/content/model/elements/factories';
 import { switchType } from 'components/editing/elements/commands/toggleTextTypes';
@@ -15,9 +15,7 @@ export const insertCodeblock = createButtonCommandDesc({
     if (!editor.selection) return;
     Transforms.insertNodes(editor, Model.code(), { at: editor.selection });
   },
-  precondition: (editor) => {
-    return !isActive(editor, 'table');
-  },
+  precondition: (editor) => !isActive(editor, 'table'),
 });
 
 export const toggleCodeblock = createButtonCommandDesc({
