+++ assets/src/components/editing/elements/heading/headingActions.ts
@@ -1,7 +1,7 @@
 import { createButtonCommandDesc } from 'components/editing/elements/commands/commandFactories';
 import { CommandDescription } from 'components/editing/elements/commands/interfaces';
 import { switchType } from 'components/editing/elements/commands/toggleTextTypes';
-import { getNearestBlock, isActive, isTopLevel } from 'components/editing/utils';
+import { getNearestBlock, isActive, isTopLevel } from 'components/editing/slateUtils';
 import { Editor, Element, Transforms } from 'slate';
 
 export const toggleHeading = createButtonCommandDesc({
