+++ assets/src/components/editing/elements/image/ImageElement.tsx
@@ -4,18 +4,12 @@ import * as ContentModel from 'data/content/model/elements/types';
 import { EditorProps } from 'components/editing/elements/interfaces';
 import { HoverContainer } from 'components/editing/toolbar/HoverContainer';
 import { Resizable } from 'components/misc/resizable/Resizable';
-import { Toolbar } from 'components/editing/toolbar/Toolbar';
 import { Placeholder } from 'components/editing/elements/common/Placeholder';
-import { CommandButton } from 'components/editing/toolbar/buttons/CommandButton';
 import { CaptionEditor } from 'components/editing/elements/common/settings/CaptionEditor';
 import { useElementSelected } from 'data/content/utils';
 import { selectImage } from 'components/editing/elements/image/imageActions';
 import { Maybe } from 'tsmonad';
-import { DescriptiveButton } from 'components/editing/toolbar/buttons/DescriptiveButton';
-import { modalActions } from 'actions/modal';
-import { ImageModal } from 'components/editing/elements/image/ImageModal';
-import { CommandContext } from 'components/editing/elements/commands/interfaces';
-import { createButtonCommandDesc } from 'components/editing/elements/commands/commandFactories';
+import { ImageSettings } from 'components/editing/elements/image/imageSettings';
 
 interface Props extends EditorProps<ContentModel.Image> {}
 export const ImageEditor = (props: Props) => {
@@ -58,7 +52,11 @@ export const ImageEditor = (props: Props) => {
         align="start"
         position="top"
         content={
-          <Settings model={props.model} onEdit={onEdit} commandContext={props.commandContext} />
+          <ImageSettings
+            model={props.model}
+            onEdit={onEdit}
+            commandContext={props.commandContext}
+          />
         }
       >
         <div>
@@ -75,63 +73,3 @@ export const ImageEditor = (props: Props) => {
     </div>
   );
 };
-
-interface SettingsProps {
-  commandContext: CommandContext;
-  model: ContentModel.Image;
-  onEdit: (attrs: Partial<ContentModel.Image>) => void;
-}
-const Settings = (props: SettingsProps) => {
-  return (
-    <Toolbar context={props.commandContext}>
-      <Toolbar.Group>
-        <SelectImageButton model={props.model} onEdit={props.onEdit} />
-      </Toolbar.Group>
-      <Toolbar.Group>
-        <SettingsButton model={props.model} onEdit={props.onEdit} />
-      </Toolbar.Group>
-    </Toolbar>
-  );
-};
-interface SelectImageProps {
-  model: ContentModel.Image;
-  onEdit: (attrs: Partial<ContentModel.Image>) => void;
-}
-const SelectImageButton = (props: SelectImageProps) => (
-  <CommandButton
-    description={createButtonCommandDesc({
-      icon: 'insert_photo',
-      description: 'Select Image',
-      execute: (context, _editor) =>
-        selectImage(context.projectSlug, props.model.src).then((selection) =>
-          Maybe.maybe(selection).map((src) => props.onEdit({ src })),
-        ),
-    })}
-  />
-);
-
-interface SettingsButtonProps {
-  model: ContentModel.Image;
-  onEdit: (attrs: Partial<ContentModel.Image>) => void;
-}
-const SettingsButton = (props: SettingsButtonProps) => (
-  <DescriptiveButton
-    description={createButtonCommandDesc({
-      icon: '',
-      description: 'Settings',
-      execute: (_context, _editor, _params) =>
-        window.oliDispatch(
-          modalActions.display(
-            <ImageModal
-              model={props.model}
-              onDone={({ alt, width }: Partial<ContentModel.Image>) => {
-                window.oliDispatch(modalActions.dismiss());
-                props.onEdit({ alt, width });
-              }}
-              onCancel={() => window.oliDispatch(modalActions.dismiss())}
-            />,
-          ),
-        ),
-    })}
-  />
-);
