+++ assets/src/components/editing/elements/link/LinkCmd.tsx
@@ -1,9 +1,11 @@
 import { Transforms, Editor, Element } from 'slate';
-import { Command, CommandDescription } from 'components/editing/elements/commands/interfaces';
-import { isActive } from '../../utils';
+import { isActive } from '../../slateUtils';
 import { Model } from 'data/content/model/elements/factories';
+import { createButtonCommandDesc } from '../commands/commandFactories';
 
-const command: Command = {
+export const commandDesc = createButtonCommandDesc({
+  icon: 'insert_link',
+  description: 'Link (⌘L)',
   execute: (_context, editor, _params) => {
     const selection = editor.selection;
     if (!selection) return;
@@ -17,15 +19,6 @@ const command: Command = {
     const href = Editor.string(editor, selection);
     Transforms.wrapNodes(editor, Model.link(href), { split: true });
   },
-  precondition: (editor) => {
-    return !isActive(editor, ['code']);
-  },
-};
-
-export const commandDesc: CommandDescription = {
-  type: 'CommandDesc',
-  icon: () => 'insert_link',
-  description: () => 'Link (⌘L)',
-  command,
+  precondition: (editor) => !isActive(editor, ['code']),
   active: (e) => isActive(e, 'a'),
-};
+});
