+++ assets/src/components/editing/elements/marks/toggleMarkActions.ts
@@ -1,5 +1,5 @@
 import { popupCmdDesc } from 'components/editing/elements/popup/PopupCmd';
-import { isActive, isMarkActive } from 'components/editing/utils';
+import { isActive, isMarkActive } from 'components/editing/slateUtils';
 import { Command } from 'components/editing/elements/commands/interfaces';
 import { Mark } from 'data/content/model/text';
 import { Editor } from 'slate';
