+++ assets/src/components/editing/elements/paragraph/paragraphActions.ts
@@ -4,7 +4,7 @@ import { createButtonCommandDesc } from 'components/editing/elements/commands/co
 import { switchType } from 'components/editing/elements/commands/toggleTextTypes';
 import { toggleHeading } from 'components/editing/elements/heading/headingActions';
 import { toggleUnorderedList } from 'components/editing/elements/list/listActions';
-import { isActive } from 'components/editing/utils';
+import { isActive } from 'components/editing/slateUtils';
 
 export const toggleParagraph = createButtonCommandDesc({
   icon: 'subject',
