+++ assets/src/components/editing/elements/table/commands/insertTable.tsx
@@ -1,6 +1,6 @@
 import { Editor, Transforms } from 'slate';
 import { Model } from 'data/content/model/elements/factories';
-import { isTopLevel } from 'components/editing/utils';
+import { isTopLevel } from 'components/editing/slateUtils';
 import { createButtonCommandDesc } from 'components/editing/elements/commands/commandFactories';
 
 export const insertTable = createButtonCommandDesc({
