+++ assets/src/components/editing/toolbar/buttons/CommandButton.tsx
@@ -1,6 +1,6 @@
 import { CommandDescription } from 'components/editing/elements/commands/interfaces';
 import { ButtonContent } from 'components/editing/toolbar/buttons/shared';
-import { useToolbar } from 'components/editing/toolbar/useToolbar';
+import { useToolbar } from 'components/editing/toolbar/hooks/useToolbar';
 import React from 'react';
 import { useSlate } from 'slate-react';
 import { classNames } from 'utils/classNames';
