+++ assets/src/components/editing/toolbar/buttons/DescriptiveButton.tsx
@@ -1,5 +1,5 @@
 import { CommandDescription } from 'components/editing/elements/commands/interfaces';
-import { useToolbar } from 'components/editing/toolbar/useToolbar';
+import { useToolbar } from 'components/editing/toolbar/hooks/useToolbar';
 import React from 'react';
 import { useSlate } from 'slate-react';
 import { classNames } from 'utils/classNames';
