+++ assets/src/components/editing/toolbar/buttons/DropdownButton.tsx
@@ -1,6 +1,6 @@
 import { CommandDescription } from 'components/editing/elements/commands/interfaces';
 import { ButtonContent } from 'components/editing/toolbar/buttons/shared';
-import { useToolbar } from 'components/editing/toolbar/useToolbar';
+import { useToolbar } from 'components/editing/toolbar/hooks/useToolbar';
 import React, { PropsWithChildren } from 'react';
 import { Popover } from 'react-tiny-popover';
 import { useSlate } from 'slate-react';
