+++ assets/src/components/resource/editors/dragndrop/handlers/dragStart.ts
@@ -1,6 +1,6 @@
 import { DragPayload } from '../interfaces';
 import { ResourceContent } from 'data/content/resource';
-import { toSimpleText } from 'components/editing/utils';
+import { toSimpleText } from 'components/editing/slateUtils';
 
 export const dragStartHandler =
   (
